Move wiser switches of kind switch to the correct entity domain
